backend/requirements.txt
itsdangerous==1.1.0
SpeechRecognition==3.8.1
typing==3.6.6
zappa-sentry==0.2.3
pyasn1==0.4.4
itypes==1.1.0
djangorestframework==3.9.0
PySocks==1.6.8
rsa==3.4.2
django-ses==0.8.7
django-s3-storage==0.12.4
boto==2.49.0
PyYAML==3.13
mysqlclient==1.3.13
requests==2.20.1
oauth2client==4.1.3
Unidecode==1.0.22
six==1.11.0
coreschema==0.0.4
httplib2==0.11.3
awscli==1.16.47
cachetools==2.1.0
kappa==0.6.0
django-phonenumber-field==2.1.0
django-storages==1.7.1
asn1crypto==0.24.0
PyJWT==1.6.4
# use the backup branch -e git+https://github.com/pnija/zappa.git@bkp-0.48.2#egg=zappa.
-e git+https://github.com/jnoortheen/zappa.git@4ad3977a58aa2c7a1474b45e2e0ea726afc0244e#egg=zappa
-e git+https://github.com/Miserlou/lambda-packages.git@d436c0b6c7582e0ca9ede7b05204d8f913eb3942#egg=lambda_packages
jsonpickle==0.9.6
djangorestframework-jwt==1.11.0
cryptography==2.3.1
idna==2.7
pynamodb==3.3.1
s3transfer==0.1.13
colorama==0.3.9
django-dbbackup==3.2.0
Werkzeug==0.14.1
django-utils-plus==0.6.7
twilio==6.19.2
google-api-python-client==1.7.4
google-auth==1.5.1
wsgi-request-logger==0.4.6
pytz==2018.7
botocore==1.12.37
durationpy==0.5
MarkupSafe==1.1.1
django-twilio==0.9.3
python-slugify==1.2.4
pyOpenSSL==18.0.0
django-filter==2.0.0
raven==6.9.0
django-mysql==2.4.1
hjson==3.0.1
troposphere==2.3.3
Babel==2.6.0
raven-python-lambda==0.1.10
vobject==0.9.6.1
placebo==0.8.2
google-auth-httplib2==0.0.3
python-dateutil==2.6.1
argcomplete==1.9.3
certifi==2018.10.15
future==0.16.0
tqdm==4.19.1
chardet==3.0.4
pyasn1-modules==0.2.2
Jinja2==2.10
coreapi==2.3.3
psutil==5.4.8
uritemplate==3.0.0
Click==7.0
urllib3==1.23
ndg-httpsclient==0.5.1
cffi==1.11.5
Django==1.11.16
toml==0.10.0
cfn-flip==1.0.3
docutils==0.14
django-extensions==2.1.3
pycparser==2.19
jmespath==0.9.3
pusher==2.0.2
proofread==0.1.2
django-cors-headers==2.4.0
boto3==1.9.37
chargebee==2.6.0
phonenumberslite==8.9.16
celery==4.1.0
protobuf==3.6.1
billiard==3.5.0.4
django-celery-results==1.0.1
dialogflow==0.5.1
anyjson==0.3.3
google-api-core==1.3.0
datefinder==0.6.1
grpcio==1.14.1
kombu==4.2.1
vine==1.1.4
googleapis-common-protos==1.5.3
base58==1.0.0
regex==2016.1.10
amqp==2.3.2
fake-factory==9999.9.9
configobj==5.0.6
repo==0.1.0
url==0.4.2
